requirements.txt
flask==3.0.0
requests==2.31.0
pyyaml==6.0.1
lxml==5.1.0
